Move all capture code to example
This avoids a new package, and the code is small enough that
it fits well as an example.
