+++ examples/example-keyboard-capture/package.json
@@ -7,7 +7,6 @@
     "clean": "rimraf build"
   },
   "dependencies": {
-    "@lumino/keyboard-capture": "^2.0.0",
     "@lumino/signaling": "^2.1.3",
     "@lumino/widgets": "^2.6.0"
   },
