+++ examples/example-keyboard-capture/src/capture.ts
