--- packages/keyboard-capture/package.json
@@ -1,67 +1,0 @@
-{
-  "name": "@lumino/keyboard-capture",
-  "version": "2.0.0",
-  "description": "Lumino Keyboard Capture widget",
-  "homepage": "https://github.com/jupyterlab/lumino",
-  "bugs": {
-    "url": "https://github.com/jupyterlab/lumino/issues"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/jupyterlab/lumino.git"
-  },
-  "license": "BSD-3-Clause",
-  "author": "Project Jupyter",
-  "main": "dist/index.js",
-  "jsdelivr": "dist/index.min.js",
-  "unpkg": "dist/index.min.js",
-  "module": "dist/index.es6",
-  "types": "types/index.d.ts",
-  "files": [
-    "dist/*",
-    "src/*",
-    "types/*"
-  ],
-  "scripts": {
-    "api": "api-extractor run --local --verbose",
-    "build": "yarn build:src && rollup -c",
-    "build:src": "tsc --build",
-    "build:test": "yarn clean:test && tsc --build tests && cd tests && rollup -c",
-    "clean": "rimraf ./lib *.tsbuildinfo ./types ./dist",
-    "clean:test": "rimraf tests/lib tests/tsconfig.tsbuildinfo",
-    "minimize": "terser dist/index.js -c -m --source-map \"content='dist/index.js.map',url='index.min.js.map'\" -o dist/index.min.js",
-    "test": "yarn build:test && web-test-runner tests/lib/bundle.test.js --node-resolve --playwright",
-    "test:chromium": "yarn test -- --browsers chromium",
-    "test:debug": "yarn test -- --manual --open",
-    "test:firefox": "yarn test -- --browsers firefox",
-    "test:webkit": "yarn test -- --browsers webkit",
-    "watch": "tsc --build --watch"
-  },
-  "dependencies": {
-    "@lumino/keyboard": "^2.0.2",
-    "@lumino/signaling": "^2.1.3",
-    "@lumino/widgets": "^2.6.0"
-  },
-  "devDependencies": {
-    "@lumino/messaging": "^2.0.2",
-    "@microsoft/api-extractor": "^7.36.0",
-    "@rollup/plugin-commonjs": "^24.0.0",
-    "@rollup/plugin-node-resolve": "^15.0.1",
-    "@types/chai": "^3.4.35",
-    "@types/mocha": "^2.2.39",
-    "@web/test-runner": "^0.18.2",
-    "@web/test-runner-playwright": "^0.11.0",
-    "chai": "^4.3.4",
-    "mocha": "^9.0.3",
-    "rimraf": "^5.0.1",
-    "rollup": "^3.25.1",
-    "rollup-plugin-postcss": "^4.0.2",
-    "rollup-plugin-sourcemaps": "^0.6.3",
-    "terser": "^5.18.1",
-    "tslib": "^2.3.0",
-    "typescript": "~5.1.3"
-  },
-  "publishConfig": {
-    "access": "public"
-  }
-}
