--- packages/keyboard-capture/tests/rollup.config.mjs
@@ -1,8 +1,0 @@
-/*
- * Copyright (c) Jupyter Development Team.
- * Distributed under the terms of the Modified BSD License.
- */
-
-import { createRollupTestConfig } from '@lumino/buildutils';
-const rollupConfig = createRollupTestConfig();
-export default rollupConfig;
