--- packages/keyboard-capture/tests/tsconfig.json
@@ -1,10 +1,0 @@
-{
-  "extends": "../../../tsconfigbase",
-  "compilerOptions": {
-    "lib": ["DOM", "ES2018"],
-    "outDir": "lib",
-    "rootDir": "src",
-    "types": ["chai", "mocha"]
-  },
-  "include": ["src/*"]
-}
