--- packages/keyboard-capture/tsconfig.json
@@ -1,21 +1,0 @@
-{
-  "extends": "../../tsconfigbase",
-  "compilerOptions": {
-    "declarationDir": "types",
-    "lib": ["DOM", "ES2018"],
-    "outDir": "lib",
-    "rootDir": "src"
-  },
-  "include": ["src/*"],
-  "references": [
-    {
-      "path": "../keyboard"
-    },
-    {
-      "path": "../signaling"
-    },
-    {
-      "path": "../widgets"
-    }
-  ]
-}
